#4: Move helpers to /Components
