+++ backend/Components/required_keys.js
@@ -1,0 +1,7 @@
+const required_keys = [
+    "access_token", 
+    "command", 
+    "execute_at"
+]
+
+module.required_keys = required_keys
